+++ synch/enums.py
@@ -14,3 +14,4 @@ class SourceDatabase(str, Enum):
 class ClickHouseEngine(str, Enum):
     merge_tree = "MergeTree"
     collapsing_merge_tree = "CollapsingMergeTree"
+    versioned_collapsing_merge_tree = "VersionedCollapsingMergeTree"